Update and rename 不知道 to 字母统计
The frequency of the 26 letters in English is different in practice. The frequency of the most common letter "e" is as high as 12.702%, and the frequency of the letter "t" is also 9.056%. In contrast, the frequency of the letter "q" is only 0.095%, and the frequency of the letter "z" is only 0.074%. Please write a program: input a string of English, and output the number of letters with the most repeated times.
Note: the input English is in lowercase letters and does not contain other special characters.